mypy-requirements.txt
aiohttp==3.9.0b0
aiosignal==1.3.1
anyio==4.0.0
async-timeout==4.0.3
attrs==23.1.0
black==23.10.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
coverage==7.3.2
cryptography==41.0.5
dnspython==2.4.2
frozenlist==1.4.0
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
idna==3.4
iniconfig==2.0.0
loguru==0.7.2
multidict==6.0.4
mypy==1.6.1
mypy-extensions==1.0.0
numpy==1.26.0
openai==0.28.1
packaging==23.2
pandas==2.1.1
pandas-stubs==2.1.1.230928
pathspec==0.11.2
pinecone-client==2.2.4
platformdirs==3.11.0
pluggy==1.3.0
pycparser==2.21
pymongo==4.5.0
pytest==7.4.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-telegram-bot==20.6
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
scipy==1.11.3
six==1.16.0
sniffio==1.3.0
tiktoken==0.5.1
tornado==6.3.3
tqdm==4.66.1
types-cachetools==5.3.0.7
types-colorama==0.4.15.12
types-docutils==0.20.0.3
types-openpyxl==3.1.0.26
types-Pygments==2.16.0.0
types-pyOpenSSL==23.3.0.0
types-pytz==2023.3.1.1
types-regex==2023.10.3.0
types-requests==2.31.0.10
types-setuptools==68.2.0.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.6
win32-setctime==1.1.0
yarl==1.9.2
